feat: change stockfish move fetch method from GET to POST
The fetch method for the stockfish move has been modified from GET to POST. This change was implemented to allow for more secure and robust data exchange within the endpoint. Now, the data is sent in the body of the request instead of being appended in the URL as query parameters.